requirements.txt
flask
werkzeug
python-magic
yara-python

